não permiti deletar plano com detalhes

+++ elpisfood/app/Http/Controllers/Admin/PlanController.php
@@ -40,11 +40,20 @@ class PlanController extends Controller
     }
     public function destroy($url)
     {
-        $plan = $this->repository->where('url',$url)->first();
+        $plan = $this->repository
+                    ->with('details')
+                    ->where('url',$url)
+                    ->first();
 
         if(!$plan)
         {
             return redirect()->back();
+        }
+        if($plan->details->count() > 0)
+        {
+            return redirect()
+                            ->back()
+                            ->with('error','Existem detalhes vinculados ao plano. Apague o detalhe!');
         }
         $plan->delete();
